feat: add class detection util and refactor client's item registration
This commit adds a new utility function 'isClass' in 'src/lib/utils.ts' which
checks if a function is a class. It also refactors the item registration in
'src/structures/Client.ts' to use the new 'isClass' function for validation before
instantiating module items. This update improves the robustness of the item
registration process.