feat(database): add support for PostgreSQL
This commit introduces support for PostgreSQL in the application. It
allows the database property in the abstract class 'DatabaseA' to
either be an instance of 'Database' or 'Pool'. It also modifies the
'Postgresql' implementation to use the 'Pool' class for connecting to
the database.

The 'DatabaseManager' has been updated to handle PostgreSQL connection
parameters and the SSL configuration has been added.

Finally, minor changes have been made in the 'Client' file to reflect
these changes.